requirements.txt
# NSFWGuard - Python dependencies
# Install with: pip install -r requirements.txt

# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9          # needed for Form() and UploadFile in FastAPI
jinja2==3.1.4                    # HTML templates
itsdangerous==2.2.0              # needed by SessionMiddleware
starlette==0.37.2

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.36
pymysql==1.1.1                   # MySQL driver
cryptography==42.0.7             # needed by pymysql for auth

# ── Auth ───────────────────────────────────────────────────────────────────────
passlib[bcrypt]==1.7.4           # password hashing
python-jose[cryptography]==3.3.0 # JWT tokens
httpx==0.27.0                    # async HTTP client for Google OAuth

# ── NSFW Classification ────────────────────────────────────────────────────────
nudenet==3.4.2                   # NudeNet: NSFW detector (90%+ accuracy)
opencv-python-headless==4.10.0.84 # video frame processing (headless = no GUI)
Pillow>=10.4.0                   # image loading

# ── YouTube Download ───────────────────────────────────────────────────────────
yt-dlp==2024.5.27                # download YouTube videos for classification

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1             # load .env files
numpy>=1.26.4                    # array ops for image processing
